Add react router library and create a greating component
