+++ package.json
@@ -12,7 +12,8 @@
     "babel-loader": "^9.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.1.3"
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.18.0"
   },
   "devDependencies": {
     "stylelint": "^13.13.1",
